Split tests in groups (#3785)
