+++ package.json
@@ -9,14 +9,16 @@
   },
   "scripts": {
     "format": "prettier --write \"{lib,bin,scripts,test}/**/*.js\"",
-    "debug:test": "mocha --inspect-brk --exit -t 0 test/index.js",
+    "debug:test": "mocha --inspect-brk --exit -t 0 test/all-tests-suite.js",
     "debug:tape": "node --inspect-brk test/tape/index.js",
     "coveralls": "nyc report --reporter=text-lcov | coveralls",
     "lint": "eslint \"lib/**/*.js\" \"test/**/*.js\"",
     "lint:types": "dtslint types",
     "lint:everything": "npm run lint:types && npm run lint",
-    "test": "mocha --exit -t 10000 test/index.js && npm run test:tape && npm run test:cli",
-    "test:coverage": "nyc mocha --exit --check-leaks --globals __core-js_shared__ -t 10000 test/index.js && npm run test:tape && npm run test:cli",
+    "test:unit": "mocha --exit -t 10000 test/db-less-test-suite.js && npm run test:tape && npm run test:cli",
+    "test:db": "mocha --exit -t 10000 test/integration-test-suite.js",
+    "test": "mocha --exit -t 10000 test/all-tests-suite.js && npm run test:tape && npm run test:cli",
+    "test:coverage": "nyc mocha --exit --check-leaks --globals __core-js_shared__ -t 10000 test/all-tests-suite.js && npm run test:tape && npm run test:cli",
     "test:everything": "npm run lint:everything && npm run test:coverage",
     "test:sqlite": "cross-env DB=sqlite3 npm test",
     "test:tape": "node test/tape/index.js | tap-spec",
