+++ test/all-tests-suite.js
@@ -1,0 +1,2 @@
+require('./db-less-test-suite');
+require('./integration-test-suite');
