+++ test/integration-test-suite.js
@@ -1,0 +1,8 @@
+const { initTests } = require('./testInitializer');
+
+initTests();
+
+describe('Integration Tests', function() {
+  this.timeout(process.env.KNEX_TEST_TIMEOUT || 5000);
+  require('./integration');
+});
