+++ test/integration/builder/inserts.js
@@ -6,7 +6,7 @@ const uuid = require('uuid');
 const _ = require('lodash');
 const sinon = require('sinon');
 
-const { TEST_TIMESTAMP } = require('../../constants');
+const { TEST_TIMESTAMP } = require('../../util/constants');
 
 module.exports = function(knex) {
   describe('Inserts', function() {
