+++ test/integration/builder/joins.js
@@ -2,7 +2,7 @@
 
 const { expect } = require('chai');
 
-const { TEST_TIMESTAMP } = require('../../constants');
+const { TEST_TIMESTAMP } = require('../../util/constants');
 
 module.exports = function(knex) {
   describe('Joins', function() {
