+++ test/integration/builder/selects.js
@@ -6,7 +6,7 @@ const _ = require('lodash');
 const assert = require('assert');
 const Runner = require('../../../lib/runner');
 
-const { TEST_TIMESTAMP } = require('../../constants');
+const { TEST_TIMESTAMP } = require('../../util/constants');
 
 module.exports = function(knex) {
   describe('Selects', function() {
