+++ test/integration/builder/updates.js
@@ -2,7 +2,7 @@
 
 const { expect } = require('chai');
 
-const { TEST_TIMESTAMP } = require('../../constants');
+const { TEST_TIMESTAMP } = require('../../util/constants');
 
 module.exports = function(knex) {
   describe('Updates', function() {
